backend/pyproject.toml
[project]
name = "devforge-backend"
version = "0.1.0"
description = "DevForge backend service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.32",
  "asyncpg>=0.29.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.4",
  "httpx>=0.27.0",
  "redis>=5.0.7",
  "celery[redis]>=5.4.0",
  "python-jose[cryptography]>=3.3.0",
  "cryptography>=43.0.0",
  "passlib[bcrypt]>=1.7.4",
  "structlog>=24.4.0",
  "prometheus-client>=0.20.0",
  "tenacity>=8.5.0",
  "python-multipart>=0.0.9",
  "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.5.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
